+++ CHANGELOG.md
@@ -19,6 +19,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Deprecated `include` and `exclude` parameters in `Bucket.query` and `ReplicationSettings`, [PR-132](https://github.com/reductstore/reduct-py/pull/132)
 
+
+### Fixed
+
+- `head` flag in `Bucket.query` method, [PR-133](https://github.com/reductstore/reduct-py/pull/133)
+
 ## [1.15.0] - 2025-05-05
 
 ### Added